fix(chat): title a session whose first message is a slash command (#257)
Lifts the title-assignment block out of send_message's plain-text tail
into a shared assign_initial_title helper, and splits send_message into
dispatch_command and send_plain_text so the command branch's early
return no longer skips titling. New module: chat_manager/send.rs.

cargo test -p mainframe-chat: 192 passed; 0 failed (was 187 passed, 5
failed before this commit — the five Group A red-phase tests are now
green: provider_command_first_message_sets_the_fallback_title,
mainframe_command_first_message_titles_from_typed_text_not_the_wrapper,
command_first_message_generated_title_overwrites_the_fallback,
title_generation_is_not_awaited_by_a_command_send,
command_first_fallback_survives_a_generation_that_returns_nothing.
cargo clippy --all-targets -- -D warnings: clean. cargo fmt --check: clean.

Sizes: chat_manager.rs 2351 -> 2157; send.rs 290 (new); every function
in send.rs is under 50 lines. send_message keeps its 91-line preamble
(now ~99 lines total), a pre-existing violation deferred to #292 per
the plan's Constraints section, alongside chat_manager.rs staying over
the 300-line file limit.
